security update:remove branch name injection
Branch names are external user generated, therefore, they constitute a vulnerability point (altought very niche). For this reason it's better to limit user generated input in shell scripts like the one in our file, when using clever-tools. Instead of using env.BRANCH_NAME in our script, we'll use github.event.number, which is the number of the PR on the repo.